Process files with multiple ids or that do not follow naming schema
This change will allow to process multiple entries on the inventory that are recorded on a single file (e.g. physical media that has multiple labels on a single side).

This will also add files that do not follow the defined naming shcema to the exceptions worksheet.

procdigs is the program that checks for multiple ids on a single file. This makes procdigs aware of the naming schema. This should be changed at some point to make it agnostic and have it only read intrinsic and extrinsic file details.